+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-Version 7.6.0
+Version 7.8.0
 
 # Changelog
 
@@ -7,7 +7,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased - yyyy-mm-dd
+## 7.8.0 - 2025-02-19
 
 ### Fixed